All variables, constants, and parameters now explicitly double precision - real precision is not accurate enough.

+++ derivs.f90
@@ -14,8 +14,8 @@ module derivs
     use parameters
     implicit none
 
-    real, intent(in)  :: f(0:nx,0:nz)
-    real, intent(out) :: fx(0:nx,0:nz)
+    double precision, intent(in)  :: f(0:nx,0:nz)
+    double precision, intent(out) :: fx(0:nx,0:nz)
     integer :: j, k
 
     do j = 1, nx1
@@ -30,12 +30,12 @@ module derivs
     use parameters
     implicit none
 
-    real, intent(in)  :: f(0:nx,0:nz)
-    real, intent(out) :: fxx(0:nx,0:nz)
+    double precision, intent(in)  :: f(0:nx,0:nz)
+    double precision, intent(out) :: fxx(0:nx,0:nz)
     integer :: j, k
 
     do j = 1, nx1
-      fxx(j,:) = f(j+1,:) - 2.0 * f(j,:) + f(j-1,:)
+      fxx(j,:) = f(j+1,:) - 2d0 * f(j,:) + f(j-1,:)
     end do
 
     return
@@ -46,14 +46,14 @@ module derivs
     use parameters
     implicit none
 
-    real, intent(in)  :: f(0:nx,0:nz)
-    real, intent(out) :: fz(0:nx,0:nz)
+    double precision, intent(in)  :: f(0:nx,0:nz)
+    double precision, intent(out) :: fz(0:nx,0:nz)
     integer :: j, k
 
     do k = 1, nz1
       fz(1:nx1,k) = f(1:nx1,k+1) - f(1:nx1,k-1)
-      fz(1:nx1,0) = -f(1:nx1,2) + 4.0 * f(1:nx1,1) - 3.0 * f(1:nx1,0)
-      fz(1:nx1,nz) = f(1:nx1,nz-2) - 4.0 * f(1:nx1,nz1) + 3.0 * &
+      fz(1:nx1,0) = -f(1:nx1,2) + 4d0 * f(1:nx1,1) - 3d0 * f(1:nx1,0)
+      fz(1:nx1,nz) = f(1:nx1,nz-2) - 4d0 * f(1:nx1,nz1) + 3d0 * &
                      f(1:nx1,nz)   !forward/backward difference at boundaries
     end do
 
@@ -65,16 +65,16 @@ module derivs
     use parameters
     implicit none
 
-    real, intent(in)  :: f(0:nx,0:nz)
-    real, intent(out) :: fzz(0:nx,0:nz)
+    double precision, intent(in)  :: f(0:nx,0:nz)
+    double precision, intent(out) :: fzz(0:nx,0:nz)
     integer :: j, k
 
     do k = 1, nz1
-      fzz(1:nx1,k) = f(1:nx1,k+1) - 2.0 * f(1:nx1,k) + f(1:nx1,k-1)
-      fzz(1:nx1,0) = -f(1:nx1,3) + 4.0 * f(1:nx1,2) - &
-                     5.0 * f(1:nx1,1) + 2.0 * f(1:nx1,0) !forward/backward
-      fzz(1:nx1,nz) = -f(1:nx1,nz-3) + 4.0 * f(1:nx1,nz-2) - & !difference at
-                      5.0 * f(1:nx1,nz1) + 2.0 * f(1:nx1,nz)  !boundaries
+      fzz(1:nx1,k) = f(1:nx1,k+1) - 2d0 * f(1:nx1,k) + f(1:nx1,k-1)
+      fzz(1:nx1,0) = -f(1:nx1,3) + 4d0 * f(1:nx1,2) - &
+                     5d0 * f(1:nx1,1) + 2d0 * f(1:nx1,0) !forward/backward
+      fzz(1:nx1,nz) = -f(1:nx1,nz-3) + 4d0 * f(1:nx1,nz-2) - & !difference at
+                      5d0 * f(1:nx1,nz1) + 2d0 * f(1:nx1,nz)  !boundaries
     end do
 
     return
