+++ linear.f90
@@ -16,8 +16,8 @@ module linear
     use ic_bc, only : s
     implicit none
 
-    real, intent(in) :: uo(0:nx,0:nz)
-    real, intent(out) :: u(0:nx,0:nz)
+    double precision, intent(in) :: uo(0:nx,0:nz)
+    double precision, intent(out) :: u(0:nx,0:nz)
     integer :: j, k
     type (deriv) :: du
 
@@ -26,22 +26,22 @@ module linear
     call deriv_zz(uo, du%zz)
 
     do k = 1, nz1
-      u(1:nx1,k) = uo(1:nx1,k) + 0.5 * rxx * du%xx(1:nx1,k) + &
-                   one_eta * rx * 0.25 * du%x(1:nx1,k) / s(1:nx1) - &
-                   one_eta**2 * dt * 0.5 * uo(1:nx1,k) / s(1:nx1)**2 + &
-                   0.5 * rzz * du%zz(1:nx1,k)
+      u(1:nx1,k) = uo(1:nx1,k) + 0.5d0 * rxx * du%xx(1:nx1,k) + &
+                   one_eta * rx * 0.25d0 * du%x(1:nx1,k) / s(1:nx1) - &
+                   one_eta**2 * dt * 0.5d0 * uo(1:nx1,k) / s(1:nx1)**2 + &
+                   0.5d0 * rzz * du%zz(1:nx1,k)
     end do
 
-    if (abs(tau - 1.0) > epsilon(tau)) then
-      u(1:nx1,0) = uo(1:nx1,0) + 0.5 * rxx * du%xx(1:nx1,0) + &
-                   one_eta * rx * 0.25 * du%x(1:nx1,0) / s(1:nx1) - &
-                   one_eta**2 * dt * 0.5 * uo(1:nx1,0) / s(1:nx1)**2 + &
-                   0.5 * rzz * du%zz(1:nx1,0)
+    if (abs(tau - 1d0) > epsilon(tau)) then
+      u(1:nx1,0) = uo(1:nx1,0) + 0.5d0 * rxx * du%xx(1:nx1,0) + &
+                   one_eta * rx * 0.25d0 * du%x(1:nx1,0) / s(1:nx1) - &
+                   one_eta**2 * dt * 0.5d0 * uo(1:nx1,0) / s(1:nx1)**2 + &
+                   0.5d0 * rzz * du%zz(1:nx1,0)
 
-      u(1:nx1,nz) = uo(1:nx1,nz) + 0.5 * rxx * du%xx(1:nx1,nz) + &
-                    one_eta * rx * 0.25 * du%x(1:nx1,nz) / s(1:nx1) - &
-                    one_eta**2 * dt * 0.5 * uo(1:nx1,nz) / s(1:nx1)**2 + &
-                    0.5 * rzz * du%zz(1:nx1,nz)
+      u(1:nx1,nz) = uo(1:nx1,nz) + 0.5d0 * rxx * du%xx(1:nx1,nz) + &
+                    one_eta * rx * 0.25d0 * du%x(1:nx1,nz) / s(1:nx1) - &
+                    one_eta**2 * dt * 0.5d0 * uo(1:nx1,nz) / s(1:nx1)**2 + &
+                    0.5d0 * rzz * du%zz(1:nx1,nz)
     end if
 
     return
@@ -55,8 +55,8 @@ module linear
     use ic_bc, only : s
     implicit none
 
-    real, intent(in) :: zo(0:nx,0:nz)
-    real, intent(out) :: zn(0:nx,0:nz)
+    double precision, intent(in) :: zo(0:nx,0:nz)
+    double precision, intent(out) :: zn(0:nx,0:nz)
     integer :: j, k
     type (deriv) :: dz
 
@@ -65,10 +65,10 @@ module linear
     call deriv_zz(zo, dz%zz)
 
     do k = 1, nz1
-      zn(1:nx1,k) = zo(1:nx1,k) + 0.5 * rxx * dz%xx(1:nx1,k) + &
-                    one_eta * rx * 0.25 * dz%x(1:nx1,k) / s(1:nx1) - &
-                    one_eta**2 * dt * 0.5 * zo(1:nx1,k) / s(1:nx1)**2 + &
-                    0.5 * rzz * dz%zz(1:nx1,k)
+      zn(1:nx1,k) = zo(1:nx1,k) + 0.5d0 * rxx * dz%xx(1:nx1,k) + &
+                    one_eta * rx * 0.25d0 * dz%x(1:nx1,k) / s(1:nx1) - &
+                    one_eta**2 * dt * 0.5d0 * zo(1:nx1,k) / s(1:nx1)**2 + &
+                    0.5d0 * rzz * dz%zz(1:nx1,k)
     end do
 
     return
